+++ kitty/child-monitor.c
@@ -761,7 +761,7 @@ prepare_to_render_os_window(OSWindow *os_window, monotonic_t now, unsigned int *
     bool was_previously_rendered_with_layers = os_window->needs_layers;
     os_window->needs_layers = (
         !global_state.supports_framebuffer_srgb || effective_os_window_alpha(os_window) < 1.f ||
-        os_window->live_resize.in_progress || (os_window->bgimage && os_window->bgimage->texture_id > 0)
+        os_window->live_resize.in_progress || (background_image_for_os_window(os_window) != NULL)
     );
     if (TD.screen && os_window->num_tabs && !os_window->has_too_few_tabs) {
         if (!os_window->tab_bar_data_updated) {
@@ -983,7 +983,7 @@ render_os_window(OSWindow *w, monotonic_t now, bool scan_for_animated_images) {
     if (!w->fonts_data) { log_error("No fonts data found for window id: %llu", w->id); return false; }
     if (prepare_to_render_os_window(w, now, &active_window_id, &active_window_bg, &num_visible_windows, &all_windows_have_same_bg, scan_for_animated_images)) needs_render = true;
     if (w->last_active_window_id != active_window_id || w->last_active_tab != w->active_tab || w->focused_at_last_render != w->is_focused) needs_render = true;
-    if (w->render_calls < 3 && w->bgimage && w->bgimage->texture_id) needs_render = true;
+    if (w->render_calls < 3 && background_image_for_os_window(w) != NULL) needs_render = true;
     if (needs_render) render_prepared_os_window(w, active_window_id, active_window_bg, num_visible_windows, all_windows_have_same_bg);
     if (w->is_focused) change_menubar_title(w->window_title);
     return needs_render;