+++ kitty/options/parse.py
@@ -8,22 +8,23 @@ from kitty.conf.utils import (
     to_color, to_color_or_none, unit_float
 )
 from kitty.options.utils import (
-    action_alias, active_tab_title_template, allow_hyperlinks, bell_on_tab, box_drawing_scale,
-    clear_all_mouse_actions, clear_all_shortcuts, clipboard_control, clone_source_strategies,
-    config_or_absolute_path, confirm_close, copy_on_select, cursor_blink_interval, cursor_text_color,
-    cursor_trail_decay, deprecated_adjust_line_height, deprecated_hide_window_decorations_aliases,
-    deprecated_macos_show_window_title_in_menubar_alias, deprecated_scrollback_indicator_opacity,
-    deprecated_send_text, disable_ligatures, edge_width, env, filter_notification, font_features,
-    hide_window_decorations, macos_option_as_alt, macos_titlebar_color, menu_map, modify_font,
-    mouse_hide_wait, narrow_symbols, notify_on_cmd_finish, optional_edge_width, parse_font_spec,
-    parse_map, parse_mouse_map, paste_actions, pointer_shape_when_dragging, remote_control_password,
-    resize_debounce_time, scrollback_lines, scrollback_pager_history_size, scrollbar_color,
-    shell_integration, show_hyperlink_targets, store_multiple, symbol_map, tab_activity_symbol,
-    tab_bar_edge, tab_bar_margin_height, tab_bar_min_tabs, tab_fade, tab_font_style, tab_separator,
-    tab_title_template, text_fg_override_threshold, titlebar_color, to_cursor_shape,
-    to_cursor_unfocused_shape, to_font_size, to_layout_names, to_modifiers,
-    transparent_background_colors, underline_exclusion, url_prefixes, url_style, visual_bell_duration,
-    visual_window_select_characters, window_border_width, window_logo_scale, window_size
+    action_alias, active_tab_title_template, allow_hyperlinks, background_images, bell_on_tab,
+    box_drawing_scale, clear_all_mouse_actions, clear_all_shortcuts, clipboard_control,
+    clone_source_strategies, config_or_absolute_path, confirm_close, copy_on_select,
+    cursor_blink_interval, cursor_text_color, cursor_trail_decay, deprecated_adjust_line_height,
+    deprecated_hide_window_decorations_aliases, deprecated_macos_show_window_title_in_menubar_alias,
+    deprecated_scrollback_indicator_opacity, deprecated_send_text, disable_ligatures, edge_width, env,
+    filter_notification, font_features, hide_window_decorations, macos_option_as_alt,
+    macos_titlebar_color, menu_map, modify_font, mouse_hide_wait, narrow_symbols, notify_on_cmd_finish,
+    optional_edge_width, parse_font_spec, parse_map, parse_mouse_map, paste_actions,
+    pointer_shape_when_dragging, remote_control_password, resize_debounce_time, scrollback_lines,
+    scrollback_pager_history_size, scrollbar_color, shell_integration, show_hyperlink_targets,
+    store_multiple, symbol_map, tab_activity_symbol, tab_bar_edge, tab_bar_margin_height,
+    tab_bar_min_tabs, tab_fade, tab_font_style, tab_separator, tab_title_template,
+    text_fg_override_threshold, titlebar_color, to_cursor_shape, to_cursor_unfocused_shape,
+    to_font_size, to_layout_names, to_modifiers, transparent_background_colors, underline_exclusion,
+    url_prefixes, url_style, visual_bell_duration, visual_window_select_characters, window_border_width,
+    window_logo_scale, window_size
 )
 
 
@@ -77,7 +78,7 @@ class Parser:
         ans['background_blur'] = int(val)
 
     def background_image(self, val: str, ans: dict[str, typing.Any]) -> None:
-        ans['background_image'] = config_or_absolute_path(val)
+        ans['background_image'] = background_images(val)
 
     def background_image_layout(self, val: str, ans: dict[str, typing.Any]) -> None:
         val = val.lower()