+++ kitty/state.c
@@ -209,6 +209,70 @@ free_bgimage(BackgroundImage **bgimage, bool release_texture) {
     bgimage = NULL;
 }
 
+static void
+free_global_background_images(bool release_texture) {
+    if (global_state.background_images.images) {
+        for (size_t i = 0; i < global_state.background_images.count; i++)
+            free_bgimage(&global_state.background_images.images[i], release_texture);
+        free(global_state.background_images.images);
+        global_state.background_images.images = NULL;
+    }
+    zero_at_ptr(&global_state.background_images);
+}
+
+static void
+ensure_background_images_generation(bool release_texture) {
+    if (global_state.background_images.generation == OPT(background_images).generation) return;
+    free_global_background_images(release_texture);
+    global_state.background_images.generation = OPT(background_images).generation;
+    global_state.background_images.images = calloc(
+            OPT(background_images).count, sizeof(global_state.background_images.images[0]));
+    if (!global_state.background_images.images) fatal("Out of memory");
+}
+
+static unsigned bg_image_id_counter = 0;
+
+static BackgroundImage*
+global_background_image(size_t idx) {
+    ensure_background_images_generation(true);
+    while (global_state.background_images.count <= idx && OPT(background_images).count) {
+        RAII_ALLOC(char, path, OPT(background_images).paths[0]);  // transfer ownership
+        remove_i_from_array(OPT(background_images.paths), 0, OPT(background_images).count);
+        BackgroundImage *img = calloc(1, sizeof(BackgroundImage));
+        if (!img) fatal("Out of memory");
+        if (image_path_to_bitmap(path, &img->bitmap, &img->width, &img->height, &img->mmap_size)) {
+            img->refcnt++;
+            img->id = ++bg_image_id_counter;
+            global_state.background_images.images[global_state.background_images.count++] = img;
+            send_bgimage_to_gpu(OPT(background_image_layout), img);
+        } else free(img);
+    }
+    return global_state.background_images.count > idx ? global_state.background_images.images[idx] : NULL;
+}
+
+BackgroundImage*
+background_image_for_os_window(OSWindow *w) {
+    if (w->background_image.no_image) return NULL;
+    if (w->background_image.override) return w->background_image.override;
+    BackgroundImage *ans = NULL;
+    if (w->background_image.global_bg_images_idx && (
+        ans = global_background_image(w->background_image.global_bg_images_idx))) return ans;
+    return global_background_image(0);
+}
+
+static size_t
+increment_bg_image_idx(size_t idx, int delta) {
+    if (delta == 0) return idx;
+    if (delta > 0) {
+        size_t new_idx = idx + delta;
+        return global_background_image(new_idx) ? new_idx : 0;
+    }
+    if (-delta <= (ssize_t)idx) return idx + delta;
+    // wrap to last image, which means we need to load all
+    global_background_image(global_state.background_images.count + OPT(background_images).count + 1);
+    return global_state.background_images.count ? global_state.background_images.count - 1 : 0;
+}
+
 OSWindow*
 add_os_window(void) {
     WITH_OS_WINDOW_REFS
@@ -219,23 +283,6 @@ add_os_window(void) {
     ans->tab_bar_render_data.vao_idx = create_cell_vao();
     ans->background_opacity.alpha = OPT(background_opacity);
     ans->created_at = monotonic();
-
-    bool wants_bg = OPT(background_image) && OPT(background_image)[0] != 0;
-    if (wants_bg) {
-        if (!global_state.bgimage) {
-            global_state.bgimage = calloc(1, sizeof(BackgroundImage));
-            if (!global_state.bgimage) fatal("Out of memory allocating the global bg image object");
-            global_state.bgimage->refcnt++;
-            if (image_path_to_bitmap(OPT(background_image), &global_state.bgimage->bitmap, &global_state.bgimage->width, &global_state.bgimage->height, &global_state.bgimage->mmap_size)) {
-                send_bgimage_to_gpu(OPT(background_image_layout), global_state.bgimage);
-            }
-        }
-        if (global_state.bgimage->texture_id) {
-            ans->bgimage = global_state.bgimage;
-            ans->bgimage->refcnt++;
-        }
-    }
-
     END_WITH_OS_WINDOW_REFS
     return ans;
 }
@@ -496,8 +543,8 @@ destroy_os_window_item(OSWindow *w) {
     Py_CLEAR(w->window_title); Py_CLEAR(w->tab_bar_render_data.screen);
     remove_vao(w->tab_bar_render_data.vao_idx);
     free(w->tabs); w->tabs = NULL;
-    free_bgimage(&w->bgimage, true);
-    zero_at_ptr(&w->bgimage);
+    free_bgimage(&w->background_image.override, true);
+    zero_at_ptr(&w->background_image);
     if (w->indirect_output.framebuffer_id) free_framebuffer(&w->indirect_output.framebuffer_id);
 }
 
@@ -951,7 +998,7 @@ PYWRAP1(os_window_has_background_image) {
     id_type os_window_id;
     PA("K", &os_window_id);
     WITH_OS_WINDOW(os_window_id)
-        if (os_window->bgimage && os_window->bgimage->texture_id > 0) { Py_RETURN_TRUE; }
+        if (background_image_for_os_window(os_window) != NULL) Py_RETURN_TRUE;
     END_WITH_OS_WINDOW
     Py_RETURN_FALSE;
 }
@@ -1322,8 +1369,12 @@ pyset_background_image(PyObject *self UNUSED, PyObject *args, PyObject *kw) {
     PyObject *os_window_ids;
     int configured = 0;
     char *png_data = NULL; Py_ssize_t png_data_size = 0;
-    static char *kwds[] = {"path", "os_window_ids", "configured", "layout_name", "png_data", "linear", "tint", "tint_gaps", NULL};
-    if (!PyArg_ParseTupleAndKeywords(args, kw, "zO!|pOy#OOO", kwds, &path, &PyTuple_Type, &os_window_ids, &configured, &layout_name, &png_data, &png_data_size, &pylinear, &pytint, &pytint_gaps)) return NULL;
+    int global_index = -1; int is_increment = 0;
+    static char *kwds[] = {
+        "path", "os_window_ids", "configured", "layout_name", "png_data", "linear", "tint", "tint_gaps", "global_index",
+        "is_increment",
+    NULL};
+    if (!PyArg_ParseTupleAndKeywords(args, kw, "zO!|pOy#OOOip", kwds, &path, &PyTuple_Type, &os_window_ids, &configured, &layout_name, &png_data, &png_data_size, &pylinear, &pytint, &pytint_gaps, &global_index, &is_increment)) return NULL;
     size_t size;
     BackgroundImageLayout layout = PyUnicode_Check(layout_name) ? bglayout(layout_name) : OPT(background_image_layout);
     BackgroundImage *bgimage = NULL;
@@ -1341,15 +1392,22 @@ pyset_background_image(PyObject *self UNUSED, PyObject *args, PyObject *kw) {
             free(bgimage);
             return NULL;
         }
-        static uint32_t bgimage_id_counter = 0;
-        bgimage->id = ++bgimage_id_counter;
+        bgimage->id = ++bg_image_id_counter;
         send_bgimage_to_gpu(layout, bgimage);
         bgimage->refcnt++;
     }
     if (configured) {
-        free_bgimage(&global_state.bgimage, true);
-        global_state.bgimage = bgimage;
-        if (bgimage) bgimage->refcnt++;
+        if (bgimage) {
+            if (global_background_image(0)) {
+                free_bgimage(&global_state.background_images.images[0], true);
+            } else {
+                free_global_background_images(true);
+                global_state.background_images.images = calloc(1, sizeof(global_state.background_images.images[0]));
+                if (!global_state.background_images.images) fatal("Out of memory");
+            }
+            global_state.background_images.images[0] = bgimage;
+            bgimage->refcnt++;
+        } else free_global_background_images(true);
         OPT(background_image_layout) = layout;
         if (pylinear && pylinear != Py_None) convert_from_python_background_image_linear(pylinear, &global_state.opts);
         if (pytint && pytint != Py_None) convert_from_python_background_tint(pytint, &global_state.opts);
@@ -1358,11 +1416,20 @@ pyset_background_image(PyObject *self UNUSED, PyObject *args, PyObject *kw) {
     for (Py_ssize_t i = 0; i < PyTuple_GET_SIZE(os_window_ids); i++) {
         id_type os_window_id = PyLong_AsUnsignedLongLong(PyTuple_GET_ITEM(os_window_ids, i));
         WITH_OS_WINDOW(os_window_id)
+            unsigned idx = os_window->background_image.global_bg_images_idx;
             make_os_window_context_current(os_window);
-            free_bgimage(&os_window->bgimage, true);
-            os_window->bgimage = bgimage;
+            free_bgimage(&os_window->background_image.override, true);
+            zero_at_ptr(&os_window->background_image);
             os_window->render_calls = 0;
-            if (bgimage) bgimage->refcnt++;
+            if (bgimage) {
+                if (!configured) {  // configured means we use the zero index global image
+                    os_window->background_image.override = bgimage;
+                    bgimage->refcnt++;
+                }
+            } else if (is_increment) {
+                os_window->background_image.global_bg_images_idx = increment_bg_image_idx(idx, global_index);
+            } else if (global_index < 0) os_window->background_image.no_image = true;
+            else os_window->background_image.global_bg_images_idx = global_index;
         END_WITH_OS_WINDOW
     }
     if (bgimage) free_bgimage(&bgimage, true);
@@ -1700,9 +1767,6 @@ finalize(void) {
     }
     if (detached_windows.windows) free(detached_windows.windows);
     detached_windows.capacity = 0;
-#define F(x) free(OPT(x)); OPT(x) = NULL;
-    F(background_image); F(bell_path); F(bell_theme); F(default_window_logo);
-#undef F
     Py_CLEAR(global_state.options_object);
     free_animation(OPT(animation.cursor));
     free_animation(OPT(animation.visual_bell));
@@ -1710,9 +1774,8 @@ finalize(void) {
     // that freeing the texture will work during shutdown and
     // the GPU driver should take care of it when the OpenGL context is
     // destroyed.
-    free_bgimage(&global_state.bgimage, false);
+    free_global_background_images(false);
     free_window_logo_table(&global_state.all_window_logos);
-    global_state.bgimage = NULL;
     free_drag_source();
     Py_CLEAR(global_state.drop_dest.data);
     zero_at_ptr(&global_state.drop_dest);